requirements - 副本 (2).txt
Cartopy==0.18.0
matplotlib==3.3.4
nest_asyncio==1.5.1
numpy==1.22.3
pandas==1.2.4
Pillow==9.0.1
plotly==5.6.0
pyecharts==1.9.1
pyshp==2.2.0
requests==2.27.1
scipy==1.6.2
streamlit==1.8.0
